SCUMM: MI1SE: Change speech.info parsing and script offset saving
This now matches the disasm but temporarily breaks our current
audio retrieval method. I will work on that separatedly.

+++ engines/scumm/scumm.h
@@ -902,9 +902,9 @@ public:
 	int _NESStartStrip = 0;
 
 	/* MI SE injected speech */
-	int16 _currentScriptSavedForSpeechMI = 0;
-	int16 _currentScriptOffsetSavedForSpeechMI = 0;
-	int16 _currentSpeechIndexMI = 0;
+	int32 _currentScriptSavedForSpeechMI = 0;
+	int32 _currentScriptOffsetSavedForSpeechMI = 0;
+	int32 _currentSpeechIndexMI = 0;
 
 protected:
 	int _curPalIndex = 0;
